Move format.ts to design system
Because it is concerned with purely presentational matters.
