+++ ui/App/ProjectScreen.svelte
@@ -14,7 +14,6 @@
   import { copyToClipboard } from "ui/src/ipc";
   import { isMaintainer, isContributor } from "ui/src/project";
 
-  import * as format from "ui/src/format";
   import * as localPeer from "ui/src/localPeer";
   import * as modal from "ui/src/modal";
   import * as notification from "ui/src/notification";
@@ -28,7 +27,7 @@
     store,
   } from "ui/src/screen/project";
 
-  import { Button, Icon, ThreeDotsMenu } from "ui/DesignSystem";
+  import { Button, Icon, ThreeDotsMenu, format } from "ui/DesignSystem";
   import Remote from "ui/App/Remote.svelte";
 
   import ScreenLayout from "ui/App/ScreenLayout.svelte";
